@@ -3,12 +3,12 @@ module.exports = class Tools {
     /**
      * read variable length integer per https://www.matroska.org/technical/specs/index.html#EBML_ex
      * @param {Uint8Array} buffer containing input
-     * @param {Number} start position in buffer
+     * @param {Number} [start=0] position in buffer
      * @returns {*}  value / length object
      */
-    static readVint(buffer, start) {
-        start = start || 0;
-        let length;
+    static readVint (buffer, start = 0) {
+
+        let length = 1;
         for (length = 1; length <= 8; length++) {
             if (buffer[start] >= Math.pow(2, 8 - length)) {
                 break;
@@ -24,7 +24,7 @@ module.exports = class Tools {
             return null;
         }
 
-        let value = buffer[start] & (1 << (8 - length)) - 1;
+        let value = buffer[start] & (1 << 8 - length) - 1;
         for (let i = 1; i < length; i++) {
             if (i === 7) {
                 if (value >= Math.pow(2, 53 - 8) && buffer[start + 7] > 0) {
@@ -49,26 +49,28 @@ module.exports = class Tools {
      * @param {Number} value to store into buffer
      * @returns {Buffer} containing the value
      */
-    static writeVint(value) {
+    static writeVint (value) {
         if (value < 0 || value > Math.pow(2, 53)) {
-            throw new Error('Unrepresentable value: ' + value);
+            throw new Error(`Unrepresentable value: ${value}`);
         }
 
-        let length;
+        let length = 1;
         for (length = 1; length <= 8; length++) {
             if (value < Math.pow(2, 7 * length) - 1) {
                 break;
             }
         }
 
-        const buffer = new Buffer(length);
+        const buffer = Buffer.alloc(length);
+        let val = value;
         for (let i = 1; i <= length; i++) {
-            const b = value & 0xFF;
+            const b = val & 0xFF;
             buffer[length - i] = b;
-            value -= b;
-            value /= Math.pow(2, 8);
+            val -= b;
+            val /= Math.pow(2, 8);
         }
         buffer[0] = buffer[0] | (1 << (8 - length));
+
         return buffer;
     }
 
@@ -79,7 +81,7 @@ module.exports = class Tools {
      * @param {Uint8Array} a2  Second array
      * @returns  {Uint8Array} concatenated arrays
      */
-    static concatenate(a1, a2) {
+    static concatenate (a1, a2) {
         if (!a1 || a1.byteLength === 0) {
             return a2;
         }
@@ -89,35 +91,33 @@ module.exports = class Tools {
         const result = new Uint8Array(a1.byteLength + a2.byteLength);
         result.set(a1, 0);
         result.set(a2, a1.byteLength);
+        // xxx: Why are these nulled out?
+        /* eslint-disable no-param-reassign */
         a1 = null;
         a2 = null;
+        /* eslint-enable no-param-reassign */
+
         return result;
     }
 
     /**
      * get a hex text string from Buff[start,end)
      * @param {Uint8Array} buff from which to read the string
-     * @param {Number} start, default 0
-     * @param {Number} end, default the whole buffer
+     * @param {Number} [start=0], default 0
+     * @param {Number} [end=buff.byteLength], default the whole buffer
      * @returns {string} the hex string
      */
-    static readHexString(buff, start, end) {
-        if (!start) {
-            start = 0;
-        }
-        if (!end) {
-            end = buff.byteLength;
-        }
-
+    static readHexString (buff, start = 0, end = buff.byteLength) {
         let result = '';
         for (let p = start; p < end; p++) {
             const q = Number(buff[p] & 0xff);
-            result += ('00' + q.toString(16)).substr(-2);
+            result += `00${q.toString(16)}`.substr(-2);
         }
+
         return result;
     }
 
-    static readUtf8(buff) {
+    static readUtf8 (buff) {
         if (typeof window === 'undefined') {
             return Buffer.from(buff).toString('utf8');
         }
@@ -127,6 +127,7 @@ module.exports = class Tools {
             if (typeof TextDecoder !== 'undefined') {
                 return new TextDecoder('utf8').decode(buff);
             }
+
             return null;
         } catch (exception) {
             return null;
@@ -138,78 +139,74 @@ module.exports = class Tools {
      * @param {Uint8Array} buff from which to read variable-length unsigned number
      * @returns {number} result (in hex for lengths > 6)
      */
-    static readUnsigned(buff) {
+    static readUnsigned (buff) {
         const b = new DataView(buff.buffer, buff.byteOffset, buff.byteLength);
         switch (buff.byteLength) {
-            case 1:
-                return b.getUint8(0);
-            case 2:
-                return b.getUint16(0);
-            case 4:
-                return b.getUint32(0);
-            default:
-                break;
+        case 1:
+            return b.getUint8(0);
+        case 2:
+            return b.getUint16(0);
+        case 4:
+            return b.getUint32(0);
+        default:
+            break;
         }
         if (buff.byteLength <= 6) {
-            let val = 0;
-            for (let i = 0; i < buff.byteLength; i++) {
-                val = (val * 256) + buff[i];
-            }
-            return val;
+            return buff.reduce((acc, current) => (acc * 256) + current, 0);
         }
 
         return Tools.readHexString(buff, 0, buff.byteLength);
 
     }
 
-    static readSigned(buff) {
+    static readSigned (buff) {
         const b = new DataView(buff.buffer, buff.byteOffset, buff.byteLength);
         switch (buff.byteLength) {
-            case 1:
-                return b.getInt8(0);
-            case 2:
-                return b.getInt16(0);
-            case 4:
-                return b.getInt32(0);
-            default:
-                return NaN;
+        case 1:
+            return b.getInt8(0);
+        case 2:
+            return b.getInt16(0);
+        case 4:
+            return b.getInt32(0);
+        default:
+            return NaN;
         }
 
     }
 
-    static readFloat(buff) {
+    static readFloat (buff) {
         const b = new DataView(buff.buffer, buff.byteOffset, buff.byteLength);
         switch (buff.byteLength) {
-            case 4:
-                return b.getFloat32(0);
-            case 8:
-                return b.getFloat64(0);
-            default:
-                return NaN;
+        case 4:
+            return b.getFloat32(0);
+        case 8:
+            return b.getFloat64(0);
+        default:
+            return NaN;
         }
     }
 
-    static readDataFromTag(tagObj, data) {
+    static readDataFromTag (tagObj, data) {
 
         tagObj.data = data;
         switch (tagObj.type) {
-            case 'u':
-                tagObj.value = Tools.readUnsigned(data);
-                break;
-            case 'f':
-                tagObj.value = Tools.readFloat(data);
-                break;
-            case 'i':
-                tagObj.value = Tools.readSigned(data);
-                break;
-            case 's':
-                tagObj.value = String.fromCharCode.apply(null, data);
-                break;
-            case '8':
-                tagObj.value = Tools.readUtf8(data);
-                break;
-            default:
-                break;
+        case 'u':
+            tagObj.value = Tools.readUnsigned(data);
+            break;
+        case 'f':
+            tagObj.value = Tools.readFloat(data);
+            break;
+        case 'i':
+            tagObj.value = Tools.readSigned(data);
+            break;
+        case 's':
+            tagObj.value = String.fromCharCode.apply(null, data);
+            break;
+        case '8':
+            tagObj.value = Tools.readUtf8(data);
+            break;
+        default:
+            break;
         }
 
         if (tagObj.name === 'SimpleBlock' || tagObj.name === 'Block') {
@@ -226,6 +223,7 @@ module.exports = class Tools {
             p += 1;
             tagObj.payload = data.subarray(p);
         }
+
         return tagObj;
     }
 };
