@@ -1,15 +1,14 @@
-const Transform = require('stream').Transform;
+const {Transform} = require('stream');
 const tools = require('./tools.js');
 const schema = require('./schema.js');
 const debug = require('debug')('ebml:decoder');
 
-const STATE_TAG = 1,
-    STATE_SIZE = 2,
-    STATE_CONTENT = 3;
+const STATE_TAG = 1;
+const STATE_SIZE = 2;
+const STATE_CONTENT = 3;
 
 module.exports = class EbmlDecoder extends Transform {
-    constructor(options) {
-        options = options || {};
+    constructor (options = {}) {
         options.readableObjectMode = true;
         super(options);
 
@@ -21,7 +20,7 @@ module.exports = class EbmlDecoder extends Transform {
         this._schema = schema;
     }
 
-    _transform(chunk, enc, done) {
+    _transform (chunk, enc, done) {
 
         if (this._buffer === null) {
             this._buffer = new Uint8Array(chunk);
@@ -44,19 +43,20 @@ module.exports = class EbmlDecoder extends Transform {
         done();
     }
 
-    getSchemaInfo(tagStr) {
+    getSchemaInfo (tagStr) {
         return this._schema[tagStr] || {
             type: 'unknown',
             name: 'unknown'
         };
     }
 
-    readTag() {
+    readTag () {
 
         debug('parsing tag');
 
         if (this._cursor >= this._buffer.length) {
             debug('waiting for more data');
+
             return false;
         }
 
@@ -65,6 +65,7 @@ module.exports = class EbmlDecoder extends Transform {
 
         if (tag == null) {
             debug('waiting for more data');
+
             return false;
         }
 
@@ -84,19 +85,20 @@ module.exports = class EbmlDecoder extends Transform {
         };
 
         this._tag_stack.push(tagObj);
-        debug('read tag: ' + tagStr);
+        debug(`read tag: ${tagStr}`);
 
         return true;
     }
 
-    readSize() {
+    readSize () {
 
         const tagObj = this._tag_stack[this._tag_stack.length - 1];
 
-        debug('parsing size for tag: ' + tagObj.tagStr);
+        debug(`parsing size for tag: ${tagObj.tagStr}`);
 
         if (this._cursor >= this._buffer.length) {
             debug('waiting for more data');
+
             return false;
         }
 
@@ -105,6 +107,7 @@ module.exports = class EbmlDecoder extends Transform {
 
         if (size == null) {
             debug('waiting for more data');
+
             return false;
         }
 
@@ -120,28 +123,30 @@ module.exports = class EbmlDecoder extends Transform {
             tagObj.end += size.value + size.length;
         }
 
-        debug('read size: ' + size.value);
+        debug(`read size: ${size.value}`);
 
         return true;
     }
 
-    readContent() {
+    readContent () {
 
         const tagObj = this._tag_stack[this._tag_stack.length - 1];
 
-        debug('parsing content for tag: ' + tagObj.tagStr);
+        debug(`parsing content for tag: ${tagObj.tagStr}`);
 
         if (tagObj.type === 'm') {
             debug('content should be tags');
             this.push(['start', tagObj]);
             this._state = STATE_TAG;
+
             return true;
         }
 
         if (this._buffer.length < this._cursor + tagObj.dataSize) {
-            debug('got: ' + this._buffer.length);
-            debug('need: ' + (this._cursor + tagObj.dataSize));
+            debug(`got: ${this._buffer.length}`);
+            debug(`need: ${this._cursor + tagObj.dataSize}`);
             debug('waiting for more data');
+
             return false;
         }
 
@@ -165,7 +170,8 @@ module.exports = class EbmlDecoder extends Transform {
             this._tag_stack.pop();
         }
 
-        debug('read data: ' + data.toString('hex'));
+        debug(`read data: ${data.toString('hex')}`);
+
         return true;
     }
 };
