@@ -7,34 +7,34 @@ const assert = require('assert');
 
 process.setMaxListeners(Infinity);
 
-describe('embl', function() {
-    describe('Values in tags', function() {
+describe('embl', () => {
+    describe('Values in tags', () => {
 
         let data = fs.readFileSync('media/video-webm-codecs-avc1-42E01E.webm');
 
-        it('should get a correct PixelWidth value from a video/webm; codecs="avc1.42E01E" file (2-byte unsigned int)', function(done) {
+        it('should get a correct PixelWidth value from a video/webm; codecs="avc1.42E01E" file (2-byte unsigned int)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'PixelWidth') {
                     assert.equal(chunk[1].value, 352);
                     done();
                 }
             });
-            decoder.on('finish', function() {
+            decoder.on('finish', () => {
                 assert.equal(0, 1, 'hit end of file without finding tag.');
                 done();
             });
             decoder.write(data);
         });
 
-        it('should get a correct EBMLVersion value from a video/webm; codecs="avc1.42E01E" file (one-byte unsigned int)', function(done) {
+        it('should get a correct EBMLVersion value from a video/webm; codecs="avc1.42E01E" file (one-byte unsigned int)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'EBMLVersion') {
                     assert.equal(chunk[1].value, 1);
                     done();
                 }
-                decoder.on('finish', function() {
+                decoder.on('finish', () => {
                     assert.equal(0, 1, 'hit end of file without finding tag.');
                     done();
                 });
@@ -42,14 +42,14 @@ describe('embl', function() {
             decoder.write(data);
         });
 
-        it('should get a correct TimeCodeScale value from a video/webm; codecs="avc1.42E01E" file (3-byte unsigned int)', function(done) {
+        it('should get a correct TimeCodeScale value from a video/webm; codecs="avc1.42E01E" file (3-byte unsigned int)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'TimecodeScale') {
                     assert.equal(chunk[1].value, 1000000);
                     done();
                 }
-                decoder.on('finish', function() {
+                decoder.on('finish', () => {
                     assert.equal(0, 1, 'hit end of file without finding tag.');
                     done();
                 });
@@ -57,29 +57,29 @@ describe('embl', function() {
             decoder.write(data);
         });
 
-        it('should get a correct TrackUID value from a video/webm; codecs="avc1.42E01E" file (56-bit integer in hex)', function(done) {
+        it('should get a correct TrackUID value from a video/webm; codecs="avc1.42E01E" file (56-bit integer in hex)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'TrackUID') {
                     assert.equal(chunk[1].value, '306d02aaa74d06');
                     done();
                 }
             });
-            decoder.on('finish', function() {
+            decoder.on('finish', () => {
                 assert.equal(0, 1, 'hit end of file without finding tag.');
                 done();
             });
             decoder.write(data);
         });
 
-        it('should get a correct DocType value from a video/webm; codecs="avc1.42E01E" file (ASCII text)', function(done) {
+        it('should get a correct DocType value from a video/webm; codecs="avc1.42E01E" file (ASCII text)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'DocType') {
                     assert.equal(chunk[1].value, 'webm');
                     done();
                 }
-                decoder.on('finish', function() {
+                decoder.on('finish', () => {
                     assert.equal(0, 1, 'hit end of file without finding tag.');
                     done();
                 });
@@ -87,14 +87,14 @@ describe('embl', function() {
             decoder.write(data);
         });
 
-        it('should get a correct MuxingApp value from a video/webm; codecs="avc1.42E01E" file (utf8 text)', function(done) {
+        it('should get a correct MuxingApp value from a video/webm; codecs="avc1.42E01E" file (utf8 text)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'MuxingApp') {
                     assert.equal(chunk[1].value, 'Chrome');
                     done();
                 }
-                decoder.on('finish', function() {
+                decoder.on('finish', () => {
                     assert.equal(0, 1, 'hit end of file without finding tag.');
                     done();
                 });
@@ -102,9 +102,9 @@ describe('embl', function() {
             decoder.write(data);
         });
 
-        it('should get a correct SimpleBlock time payload from a video/webm; codecs="avc1.42E01E" file (binary)', function(done) {
+        it('should get a correct SimpleBlock time payload from a video/webm; codecs="avc1.42E01E" file (binary)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'SimpleBlock') {
                     if (chunk[1].value > 0) {
 
@@ -116,7 +116,7 @@ describe('embl', function() {
                     }
                 }
             });
-            decoder.on('finish', function() {
+            decoder.on('finish', () => {
                 assert.equal(0, 1, 'hit end of file without finding tag.');
                 done();
             });
@@ -128,59 +128,59 @@ describe('embl', function() {
         data = fs.readFileSync('media/video-webm-codecs-vp8.webm');
 
 
-        it('should get a correct PixelWidth value from a video/webm; codecs="vp8" file (2-byte unsigned int)', function(done) {
+        it('should get a correct PixelWidth value from a video/webm; codecs="vp8" file (2-byte unsigned int)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'PixelWidth') {
                     assert.equal(chunk[1].value, 352);
                     done();
                 }
             });
-            decoder.on('finish', function() {
+            decoder.on('finish', () => {
                 assert.equal(0, 1, 'hit end of file without finding tag.');
                 done();
             });
             decoder.write(data);
         });
 
-        it('should get a correct EBMLVersion value from a video/webm; codecs="vp8" file (one-byte unsigned int)', function(done) {
+        it('should get a correct EBMLVersion value from a video/webm; codecs="vp8" file (one-byte unsigned int)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'EBMLVersion') {
                     assert.equal(chunk[1].value, 1);
                     done();
                 }
             });
-            decoder.on('finish', function() {
+            decoder.on('finish', () => {
                 assert.equal(0, 1, 'hit end of file without finding tag.');
                 done();
             });
             decoder.write(data);
         });
 
-        it('should get a correct TimeCodeScale value from a video/webm; codecs="vp8" file (3-byte unsigned int)', function(done) {
+        it('should get a correct TimeCodeScale value from a video/webm; codecs="vp8" file (3-byte unsigned int)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'TimecodeScale') {
                     assert.equal(chunk[1].value, 1000000);
                     done();
                 }
             });
-            decoder.on('finish', function() {
+            decoder.on('finish', () => {
                 assert.equal(0, 1, 'hit end of file without finding tag.');
                 done();
             });
             decoder.write(data);
         });
 
-        it('should get a correct TrackUID value from a video/webm; codecs="vp8" file (56-bit integer in hex)', function(done) {
+        it('should get a correct TrackUID value from a video/webm; codecs="vp8" file (56-bit integer in hex)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'TrackUID') {
                     assert.equal(chunk[1].value, '306d02aaa74d06');
                     done();
                 }
-                decoder.on('finish', function() {
+                decoder.on('finish', () => {
                     assert.equal(0, 1, 'hit end of file without finding tag.');
                     done();
                 });
@@ -188,39 +188,39 @@ describe('embl', function() {
             decoder.write(data);
         });
 
-        it('should get a correct DocType value from a video/webm; codecs="vp8" file (ASCII text)', function(done) {
+        it('should get a correct DocType value from a video/webm; codecs="vp8" file (ASCII text)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'DocType') {
                     assert.equal(chunk[1].value, 'webm');
                     done();
                 }
             });
-            decoder.on('finish', function() {
+            decoder.on('finish', () => {
                 assert.equal(0, 1, 'hit end of file without finding tag.');
                 done();
             });
             decoder.write(data);
         });
 
-        it('should get a correct MuxingApp value from a video/webm; codecs="vp8" file (utf8 text)', function(done) {
+        it('should get a correct MuxingApp value from a video/webm; codecs="vp8" file (utf8 text)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'MuxingApp') {
                     assert.equal(chunk[1].value, 'Chrome');
                     done();
                 }
             });
             decoder.write(data);
-            decoder.on('finish', function() {
+            decoder.on('finish', () => {
                 assert.equal(0, 1, 'hit end of file without finding tag.');
                 done();
             });
         });
 
-        it('should get a correct SimpleBlock time payload rom a video/webm; codecs="vp8" file (binary)', function(done) {
+        it('should get a correct SimpleBlock time payload rom a video/webm; codecs="vp8" file (binary)', (done) => {
             const decoder = new ebml.Decoder();
-            decoder.on('data', function(chunk) {
+            decoder.on('data', (chunk) => {
                 if (chunk[0] === 'tag' && chunk[1].name === 'SimpleBlock') {
                     if (chunk[1].value > 0) {
 
@@ -233,7 +233,7 @@ describe('embl', function() {
                     }
                 }
             });
-            decoder.on('finish', function() {
+            decoder.on('finish', () => {
                 assert.equal(0, 1, 'hit end of file without finding tag.');
                 done();
             });
